Write an initial feature spec for linking between pages with Rack::Test
So far, this does not support multiple Capybara sessions. Reloading the
UI is naive in that it only reloads the UI for the current session
(page). This might work for Rack::Test but for a JavaScript driver like
Poltergeist, where the UI might be reloaded asynchronously, the UI for
the wrong session might be reloaded.